+++ .idea/Glove-Reader-Flutter.iml
@@ -1,13 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="JAVA_MODULE" version="4">
+  <component name="FacetManager">
+    <facet type="android" name="Android">
+      <configuration>
+        <option name="GEN_FOLDER_RELATIVE_PATH_APT" value="/../../Glove-Reader-Flutter/gen" />
+        <option name="GEN_FOLDER_RELATIVE_PATH_AIDL" value="/../../Glove-Reader-Flutter/gen" />
+        <option name="MANIFEST_FILE_RELATIVE_PATH" value="/../../Glove-Reader-Flutter/AndroidManifest.xml" />
+        <option name="RES_FOLDER_RELATIVE_PATH" value="/../../Glove-Reader-Flutter/res" />
+        <option name="ASSETS_FOLDER_RELATIVE_PATH" value="/../../Glove-Reader-Flutter/assets" />
+        <option name="LIBS_FOLDER_RELATIVE_PATH" value="/../../Glove-Reader-Flutter/libs" />
+        <option name="PROGUARD_LOGS_FOLDER_RELATIVE_PATH" value="/../../Glove-Reader-Flutter/proguard_logs" />
+      </configuration>
+    </facet>
+  </component>
   <component name="NewModuleRootManager" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/gen" isTestSource="false" generated="true" />
       <excludeFolder url="file://$MODULE_DIR$/.dart_tool" />
       <excludeFolder url="file://$MODULE_DIR$/.pub" />
       <excludeFolder url="file://$MODULE_DIR$/build" />
     </content>
-    <orderEntry type="inheritedJdk" />
+    <orderEntry type="jdk" jdkName="Android API 30 Platform" jdkType="Android SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="Dart SDK" level="project" />
     <orderEntry type="library" name="Dart Packages" level="project" />
