+++ README.md
@@ -5,3 +5,11 @@ Sample Flutter App for Android and iOS that scans for Bluetooth LE devices and d
 Based on https://github.com/pauldemarco/flutter_blue/tree/master/example
 
 Used to integrate with our sign language glove translator project
+
+- FolderStructure:
+    - lib
+        - models - common classes
+        - screens - screens of the app
+        - components - common widgets
+        - services - api calls
+        - themes
