+++ lib/themes/style.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+ThemeData appTheme() {
+  return ThemeData(
+    primaryColor: Colors.white,
+    accentColor: Colors.orange,
+    hintColor: Colors.white,
+    dividerColor: Colors.white,
+    buttonColor: Colors.white,
+    scaffoldBackgroundColor: Colors.black,
+    canvasColor: Colors.black,
+  );
+}
